Swift: split store-related FFI utilities out to their own file
